Refactor HTTP plugin to reduce responsibility to maintaining the server state and issuing request hook cals
